Narrow attribute name rule
Currently, an attribute name can be any sequence of identifiers and dots. For example, the parser will except a selector like:

```
[...hi..hello]
```

This can lead to issues in the getPath function (esquery.js#L35) when the full name is split around the dots and used to index into a node. 

This pull request modifies the attribute name rule to only accept identifiers separated by exactly one dot.

The resulting array from the rule is flattened using the array reduce function. I noticed a comment further down in the grammar that seemed to indicate that the array flat() function is not supported. If it is supported, the transformation can be simplified to 
```js
return [a, ...as.reduce((acc, val) => acc.concat(val), [])].join('');
```